fix: deviceTokenの登録処理にDIする
