+++ app/src/main/java/com/example/ml_notify/MyFirebaseMessagingService.kt
@@ -1,16 +1,6 @@
 package com.example.ml_notify
 
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
-import android.os.Build
 import android.util.Log
-import androidx.core.app.NotificationCompat
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.edit
-import androidx.datastore.preferences.core.stringPreferencesKey
-import androidx.datastore.preferences.preferencesDataStore
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -20,25 +10,22 @@ import kotlinx.coroutines.launch
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
 import com.example.ml_notify.domain.TaskDataHandler
-import com.google.firebase.firestore.ktx.firestore
-import com.google.firebase.ktx.Firebase
+import com.example.ml_notify.domain.repository.FcmTokenRepository
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.map
 import javax.inject.Inject
-
-private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "app_fcm_prefs")
 
 @AndroidEntryPoint
 class MyFirebaseMessagingService : FirebaseMessagingService() {
     companion object {
-        val FCM_TOKEN_KEY = stringPreferencesKey("fcm_token")
         private val TAG = "MyFirebaseMsgService"
     }
 
     // HiltにTaskDataHandlerのインスタンスを注入する
     @Inject
     lateinit var taskDataHandler: TaskDataHandler
+
+    @Inject
+    lateinit var fcmTokenRepository: FcmTokenRepository
 
     // 子コルーチンと分離させる
     private val serviceJob = SupervisorJob()
@@ -47,8 +34,8 @@ class MyFirebaseMessagingService : FirebaseMessagingService() {
     override fun onNewToken(token: String) {
         Log.d(TAG, "Refreshed token: $token")
 
-        CoroutineScope(Dispatchers.IO).launch {
-            sendRegistrationToServer(token)
+        serviceScope.launch {
+            fcmTokenRepository.sendRegistrationToken(token)
         }
     }
 
@@ -66,44 +53,6 @@ class MyFirebaseMessagingService : FirebaseMessagingService() {
 
         remoteMessage.notification?.let {
             Log.d(TAG, "Message Notification body: ${it.body}")
-        }
-    }
-
-    // サーバー(Cloud Function)にトークンを送信する
-    private suspend fun sendRegistrationToServer(newToken: String) {
-        Log.d(TAG, "sendRegistrationTokenToServer($newToken)")
-
-         val savedToken = dataStore.data.map { preferences ->
-             preferences[FCM_TOKEN_KEY]
-         }.first()
-
-        if (savedToken == newToken) {
-            Log.d(TAG, "FCM token is unchanged, skipping save.")
-            return
-        }
-
-        val deviceData = hashMapOf("deviceToken" to newToken)
-
-        // TODO: 認証機能が実装されるまではダミーユーザーを使用
-        Firebase.firestore.collection("users/dummy-user/devices")
-            .add(deviceData)
-            .addOnSuccessListener {
-                Log.d(TAG, "FCM token successfully sent to server: $newToken")
-
-                CoroutineScope(Dispatchers.IO).launch {
-                    saveTokenToDataStore(newToken)
-                }
-            }
-            .addOnFailureListener { e ->
-                Log.e(TAG, "Error sending FCM token to server", e)
-            }
-    }
-
-    private suspend fun saveTokenToDataStore(token: String) {
-        Log.d(TAG, "Saving FCM token to DataStore: $token")
-
-        dataStore.edit { preferences ->
-            preferences[FCM_TOKEN_KEY] = token
         }
     }
