+++ app/src/main/java/com/example/ml_notify/domain/repository/FcmTokenRepository.kt
@@ -1,0 +1,5 @@
+package com.example.ml_notify.domain.repository
+
+interface FcmTokenRepository {
+    suspend fun sendRegistrationToken(newToken: String)
+}
